+++ packages/insomnia/src/ui/components/dropdowns/grpc-method-dropdown/grpc-method-dropdown.tsx
@@ -1,13 +1,7 @@
 import React, { Fragment, FunctionComponent } from 'react';
 import styled from 'styled-components';
 
-import {
-  getGrpcPathSegments,
-  getShortGrpcPath,
-  groupGrpcMethodsByPackage,
-  GrpcMethodInfo,
-  NO_PACKAGE_KEY,
-} from '../../../../common/grpc-paths';
+import type { GrpcMethodInfo, GrpcPathSegments } from '../../../../main/ipc/grpc';
 import { Dropdown } from '../../base/dropdown/dropdown';
 import { DropdownButton } from '../../base/dropdown/dropdown-button';
 import { DropdownDivider } from '../../base/dropdown/dropdown-divider';
@@ -15,6 +9,40 @@ import { DropdownItem } from '../../base/dropdown/dropdown-item';
 import { GrpcMethodTag } from '../../tags/grpc-method-tag';
 import { Button } from '../../themed-button';
 import { Tooltip } from '../../tooltip';
+
+const PROTO_PATH_REGEX = /^\/(?:(?<package>[\w.]+)\.)?(?<service>\w+)\/(?<method>\w+)$/;
+const getGrpcPathSegments = (path: string) => ({
+  packageName:PROTO_PATH_REGEX.exec(path)?.groups?.package,
+  serviceName:PROTO_PATH_REGEX.exec(path)?.groups?.service,
+  methodName:PROTO_PATH_REGEX.exec(path)?.groups?.method,
+});
+// If all segments are found, return a shorter path, otherwise the original path
+export const getShortGrpcPath = (
+  { packageName, serviceName, methodName }: GrpcPathSegments,
+  fullPath: string,
+): string => {
+  return packageName && serviceName && methodName ? `/${serviceName}/${methodName}` : fullPath;
+};
+
+export const NO_PACKAGE_KEY = 'no-package';
+
+function groupBy(list: {}[], keyGetter: (item: any) => string):Record<string, any[]> {
+  const map = new Map();
+  list.forEach(item => {
+    const key = keyGetter(item);
+    const collection = map.get(key);
+    if (!collection) {
+      map.set(key, [item]);
+    } else {
+      collection.push(item);
+    }
+  });
+  return Object.fromEntries(map);
+}
+
+export const groupGrpcMethodsByPackage = (methodInfoList: GrpcMethodInfo[]): Record<string, GrpcMethodInfo[]> => {
+  return groupBy(methodInfoList, ({ segments }) => segments.packageName || NO_PACKAGE_KEY);
+};
 
 const DropdownMethodButton = styled(Button).attrs({
   variant: 'text',
